Reworked the parser to begin addressing #2
